+++ pubspec.yaml
@@ -62,7 +62,10 @@ dependencies:
   bloc: any
   flutter_animate: ^4.5.0
   animated_text_kit: ^4.2.2
+  socket_io_client: ^3.0.1
+
 dev_dependencies:
+  mockito: ^5.4.4
   flutter_test:
     sdk: flutter
